requirements.txt
crewai>=0.28.0
crewai-tools>=0.1.0
langchain>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
langchain-community>=0.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
pandas>=2.0.0
matplotlib>=3.7.0